+++ simple_manga_downloader/SMD.py
@@ -2,7 +2,6 @@
 from .modules import Mangadex
 from .modules import Mangasee
 from .modules import Mangatown
-from .modules import Heavenmanga
 from .modules import Mangakakalot
 from .modules import Manganelo
 from .modules import Config
@@ -15,37 +14,49 @@ import html
 import os
 import imghdr
 import time
+import requests
 
 
 def main():
-    args = parser()
-    mode = args.subparser_name
-    config = Config(args.custom_cfg)
-    if not config:
-        return
-    Mangadex.lang_code = config.lang_code
+    try:
+        global CONFIG
+        parser()
+        mode = ARGS.subparser_name
+        CONFIG = Config(ARGS.custom_cfg)
+        if not CONFIG:
+            return
+        Mangadex.lang_code = CONFIG.lang_code
 
-    if mode == "down":
-        main_pipeline(args.input, args, config)
-    elif mode == "update":
-        main_pipeline(config.tracked_manga.values(), args, config)
-    elif mode == "conf":
-        conf_mode(args, config)
+        if mode == "down":
+            main_pipeline(ARGS.input)
+        elif mode == "update":
+            main_pipeline(CONFIG.tracked_manga.values())
+        elif mode == "conf":
+            conf_mode()
+        elif mode == "version":
+            version_mode()
 
-    config.save_config()
+        CONFIG.save_config()
+    except KeyboardInterrupt:
+        print("\nKeyboard Interrupt detected, stopping!")
 
 
-def site_detect(link, args, directory):
+def site_detect(link, directory, check_only=False):
     '''
     Detects the site and creates a proper manga object
     '''
+    try:
+        tracked_num = int(link)
+        if 0 < tracked_num <= len(CONFIG.tracked_manga):
+            key = list(CONFIG.tracked_manga.keys())[tracked_num - 1]
+            link = CONFIG.tracked_manga[key]
+    except ValueError:
+        pass
 
     if "mangadex.org" in link or "mangadex.cc" in link:
         site = Mangadex
     elif "mangaseeonline.us" in link:
         site = Mangasee
-    elif "heavenmanga.org" in link:
-        site = Heavenmanga
     elif "mangatown.com" in link:
         site = Mangatown
     elif "mangakakalot.com" in link:
@@ -60,7 +71,7 @@ def site_detect(link, args, directory):
         print(line)
         return False
 
-    Manga = site(link, directory)
+    Manga = site(link, directory, check_only)
 
     return Manga
 
@@ -83,54 +94,93 @@ def get_cover(Manga):
     return download_image(Manga.cover_url, Manga.session, no_ext)
 
 
-def conf_mode(args, config):
-    if args.default:
-        config.reset_config()
-    elif args.clear_tracked:
-        config.clear_tracked()
+def version_mode():
+    '''
+    Version mode of the downloader
+    '''
+    print(f"SMD v{__version__}")
 
-    if args.add:
-        for link in args.add:
+    if ARGS.version_check:
+        check = check_for_update()
+        if check is not True:
+            print("Failed to check version!")
+            print(check)
+
+
+@request_exception_handler
+def check_for_update():
+    '''
+    Checks for new versions using the PyPI API
+    '''
+
+    r = requests.get("https://pypi.org/pypi/simple-manga-downloader/json")
+    r.raise_for_status()
+    releases = r.json()["releases"]
+    lastes_version = max(releases)
+    if lastes_version > __version__:
+        date = releases[lastes_version][0]["upload_time"].split("T")[0]
+        print(f"New version available: {lastes_version} ({date})")
+    else:
+        print("No new versions found")
+    return True
+
+
+def conf_mode():
+    if ARGS.default:
+        CONFIG.reset_CONFIG()
+    elif ARGS.clear_tracked:
+        CONFIG.clear_tracked()
+
+    if ARGS.add:
+        for link in ARGS.add:
             print()
-            Manga = site_detect(link, args, config)
+            Manga = site_detect(link, CONFIG.manga_directory)
             if Manga is False:
                 continue
             title = Manga.get_main(title_return=True)
             if title is True:
-                config.add_tracked(Manga)
+                CONFIG.add_tracked(Manga)
             else:
                 print(f"{title} for:\n{link}")
-    elif args.remove:
-        config.remove_tracked(args.remove)
-    if args.list:
-        config.list_tracked(args.verbose)
-    if args.m_dir:
-        config.change_dir(args.m_dir)
-    if args.position:
-        config.change_position(args.verbose)
-    if args.paths:
-        config.print_paths()
-    if args.cover:
-        config.toogle_covers()
-    if args.lang_code:
-        config.change_lang(args.lang_code)
-    if args.list_lang:
-        config.list_lang()
+    elif ARGS.remove:
+        CONFIG.remove_tracked(ARGS.remove)
+    if ARGS.list:
+        CONFIG.list_tracked(ARGS.verbose)
+    if ARGS.m_dir:
+        CONFIG.change_dir(ARGS.m_dir)
+    if ARGS.position:
+        CONFIG.change_position(ARGS.verbose)
+    if ARGS.cover:
+        CONFIG.toogle_covers()
+    if ARGS.lang_code:
+        CONFIG.change_lang(ARGS.lang_code)
+    if ARGS.list_lang:
+        CONFIG.list_lang()
+    if ARGS.timeout is not None:
+        CONFIG.change_timeout(ARGS.timeout)
+    if ARGS.print:
+        CONFIG.print_config()
 
 
-def main_pipeline(links, args, config):
+def main_pipeline(links):
     '''
     Takes a list of manga links and does all of the required stuff
     '''
+
     try:
-        custom = args.custom_dire
+        custom = ARGS.custom_dire
         if custom:
             path = custom
         else:
-            path = config.manga_directory
+            path = CONFIG.manga_directory
     except AttributeError:
-        path = config.manga_directory
+        path = CONFIG.manga_directory
     directory = Path(path).resolve()
+
+    if hasattr(ARGS, "check_only") and ARGS.check_only:
+        check_only = True
+    else:
+        check_only = False
 
     if not links:
         print("\nNo manga to download!")
@@ -144,11 +194,11 @@ def main_pipeline(links, args, config):
     total_num_ch = 0
     found_titles = {}
     for link in links:
-        Manga = site_detect(link, args, directory)
+        Manga = site_detect(link, directory, check_only)
         if not Manga:
             continue
 
-        status = handle_manga(Manga, config.covers, args)
+        status = handle_manga(Manga)
         if status:
             ready.append(Manga)
             total_num_ch += len(Manga)
@@ -170,14 +220,18 @@ def main_pipeline(links, args, config):
         print("Found 0 chapters ready to download.")
         return
 
-    print(f"Chapters to download:")
+    print("Chapters to download:")
     for title, chapter in found_titles.items():
         print(f"{title} - {len(chapter)} chapter(s):")
         for ch in chapter:
             print(f"    Chapter {ch}")
 
     print(f"\n{total_num_ch} chapter(s) ready to download")
-    confirm = input(f"Start the download? "
+
+    if check_only:
+        return
+
+    confirm = input("Start the download? "
                     "[y to confirm/anything else to cancel]: "
                     ).lower()
     if confirm == "y":
@@ -186,7 +240,7 @@ def main_pipeline(links, args, config):
         print("Aborting!")
 
 
-def handle_manga(Manga, covers, args):
+def handle_manga(Manga):
     '''
     Handles all stuff related to the Manga
     returns True if everything fine
@@ -197,13 +251,15 @@ def handle_manga(Manga, covers, args):
               f"\n{main_status}\n{Manga.manga_link}")
         return False
 
-    if hasattr(args, "name") and args.name:
-        Manga.series_title = " ".join(args.name)
+    # Checks if a custom name was given
+    if hasattr(ARGS, "name") and ARGS.name:
+        Manga.series_title = " ".join(ARGS.name)
+
     message = f"Checking \"{Manga.series_title}\""
     line_break = make_line(message)
     print(f"\n{line_break}\n{message}\n{line_break}\n")
 
-    if covers and Manga.cover_url:
+    if CONFIG.covers and Manga.cover_url:
         cov = get_cover(Manga)
         if cov is True:
             print("-----------\n"
@@ -215,8 +271,9 @@ def handle_manga(Manga, covers, args):
                   "Failed to get cover"
                   f"\n{cov}"
                   f"\n{cov_line}\n")
+
     Manga.get_chapters()
-    filter_wanted(Manga, args)
+    filter_wanted(Manga)
 
     if not Manga.chapters:
         print("Found 0 wanted chapters")
@@ -228,7 +285,7 @@ def handle_manga(Manga, covers, args):
     return chapter_info_get(Manga)
 
 
-def filter_wanted(Manga, args):
+def filter_wanted(Manga):
     '''
     Filters the chapters dict to match the criteria
     '''
@@ -236,10 +293,10 @@ def filter_wanted(Manga, args):
     chapter_list = list(Manga.chapters)
     chapter_list.sort()
 
-    if args.subparser_name == "update":
+    if ARGS.subparser_name == "update":
         wanted = (ch for ch in chapter_list)
     else:
-        wanted = filter_selection(chapter_list, args)
+        wanted = filter_selection(chapter_list)
 
     filtered = filter_downloaded(Manga.manga_dir, wanted)
 
@@ -253,30 +310,30 @@ def make_line(string):
     return "-" * len(string)
 
 
-def filter_selection(chapter_list, args):
+def filter_selection(chapter_list):
     '''A generator that yields wanted chapters based on selection'''
 
-    if args.latest:
+    if ARGS.latest:
         try:
             yield max(chapter_list)
         except ValueError:
             pass
-    elif args.range:
-        a = args.range[0]
-        b = args.range[1]
+    elif ARGS.range:
+        a = ARGS.range[0]
+        b = ARGS.range[1]
         for ch in chapter_list:
-            if a <= ch <= b and ch not in args.exclude:
+            if a <= ch <= b and ch not in ARGS.exclude:
                 yield ch
-    elif args.selection:
-        to_get = sorted(args.selection)
+    elif ARGS.selection:
+        to_get = sorted(ARGS.selection)
         for n in to_get:
             if n.is_integer():
                 n = int(n)
-            if n in chapter_list and n not in args.exclude:
+            if n in chapter_list and n not in ARGS.exclude:
                 yield n
     else:
         for ch in chapter_list:
-            if ch not in args.exclude:
+            if ch not in ARGS.exclude:
                 yield ch
 
 
@@ -343,7 +400,8 @@ def download_image(link, session, no_ext):
     Download function, gets the image from the link, limited by wrapper
     no_ext = save target Path object with no file extension
     '''
-    content = session.get(link, stream=True, timeout=5)
+    content = session.get(link, stream=True,
+                          timeout=CONFIG.download_timeout)
 
     file_type = imghdr.what("", h=content.content)
     if not file_type:
@@ -400,19 +458,25 @@ def get_chapter(Manga, num):
     ch_dir = Manga.manga_dir / chapter_name
     ch_dir.mkdir()
 
-    name_gen = page_name_gen(title,
-                             Manga.chapters[num],
-                             chapter_name)
-    for page_name, link in name_gen:
-        no_ext = ch_dir / page_name
-        image = download_image(link, Manga.session, no_ext)
-        if image is not True:
-            print("Failed to get image, skipping to next chapter")
-            failed = True
-            shutil.rmtree(ch_dir)
-            break
-        else:
-            count += 1
+    try:
+        name_gen = page_name_gen(title,
+                                 Manga.chapters[num],
+                                 chapter_name)
+        for page_name, link in name_gen:
+            no_ext = ch_dir / page_name
+            image = download_image(link, Manga.session, no_ext)
+            if image is not True:
+                print("Failed to get image, skipping to next chapter")
+                print(image)
+                failed = True
+                count = 0
+                shutil.rmtree(ch_dir)
+                break
+            else:
+                count += 1
+    except KeyboardInterrupt:
+        shutil.rmtree(ch_dir)
+        raise KeyboardInterrupt
     return (count, failed)
 
 
@@ -447,26 +511,26 @@ def parser():
                         metavar="PATH/TO/CONFIG",
                         help="Sets a custom config to use",
                         default=None)
-    parser.add_argument("-v", "--version",
-                        action="version",
-                        version=f"SMD v{__version__}")
 
     # Sub-parsers for the config and download functionality
     subparsers = parser.add_subparsers(dest="subparser_name",
                                        metavar="mode",
                                        required=True)
     parser_conf = subparsers.add_parser("conf",
-                                        help=("Program will be in "
+                                        help=("Downloader will be in "
                                               "the config edit mode"),
                                         description="Changes the settings")
     parser_down = subparsers.add_parser("down",
-                                        help=("Program will be in "
+                                        help=("Downloader will be in "
                                               "the download mode"),
                                         description=("Downloads the given manga. "
                                                      "Supports multiple links at once."))
-    subparsers.add_parser("update",
-                          help="Program will be in the update mode",
-                          description="Download new chapters from tracked list")
+    parser_update = subparsers.add_parser("update",
+                                          help="Downloader will be in the update mode",
+                                          description="Download new chapters from tracked list")
+    parser_version = subparsers.add_parser("version",
+                                           help="Downloader will be in version mode",
+                                           description="Prints the downloader version")
 
     # Parser for download mode
     parser_down.add_argument("input", nargs="+",
@@ -543,7 +607,7 @@ def parser():
     parser_conf.add_argument("-p", "--print_conf",
                              help="Print config settings",
                              action="store_true",
-                             dest="paths")
+                             dest="print")
     parser_conf.add_argument("-c", "--covers",
                              help="Toggles the cover download setting",
                              action="store_true",
@@ -556,9 +620,27 @@ def parser():
                              help="Lists all of the mangadex language codes",
                              action="store_true",
                              dest="list_lang")
+    parser_conf.add_argument("--timeout",
+                             help="Change the download timeout",
+                             metavar="SECONDS",
+                             type=int,
+                             dest="timeout")
 
-    args = parser.parse_args()
-    return args
+    # Update options
+    parser_update.add_argument("-c", "--check",
+                               help="Only check for new chapters"
+                               "without downloading and asking for input",
+                               action="store_true",
+                               dest="check_only")
+
+    # Version options
+    parser_version.add_argument("-c", "--check",
+                                help="Checks if there is a new version available",
+                                action="store_true",
+                                dest="version_check")
+
+    global ARGS
+    ARGS = parser.parse_args()
 
 
 if __name__ == '__main__':